Fix parseURL()
The error "Notice: Trying to access array offset on value of type null" occurs when you're trying to access an element of an array using an offset (index) that doesn't exist or is null.
In the code you provided, the error is triggered in the following line: $url = $this->parseURL();

The parseURL() function is expected to return an array, but there are scenarios where it may return null. When the URL parameter ($_GET['url']) is not set, the function doesn't assign any value to $url, and as a result, it remains null.

Later in the code, you're accessing specific indices of the $url array without checking if they exist or if the array is null. For example:
$this->controller = $url[0];

If $url is null, trying to access $url[0] will trigger the error because you're attempting to access an array offset on a null value.

To resolve this issue, you can use the null coalescing operator (??) to provide a default value when the array offset doesn't exist or is null. Here's an example: $controllerName = ucfirst($url[0] ?? '');

In this case, if $url[0] is null or doesn't exist, an empty string will be assigned to $controllerName. This prevents the error from occurring and ensures that the code continues executing correctly.

By applying the null coalescing operator and implementing some additional improvements, the refined code I provided in the previous response mitigates the "array offset" error and enhances the overall structure of your PHP MVC code.